CORE-596 Extracted generative actions tests. Added case for long keys.
